src/main/java/net/skytreader/kode/chesstemplar/pieces/King.java
package net.skytreader.kode.chesstemplar.pieces;

import java.awt.Point;

import java.util.HashSet;
import java.util.Set;

import net.skytreader.kode.chesstemplar.Board;

public class King extends ChessPiece{
    
    public King(boolean isWhite){
        this.color = isWhite;
        this.pieceName = "KING";
    }

    @Override
    public Set<Point> getLegalMoves(int r, int c, Board b) throws NotMeException{
        ChessPiece indicatedPiece = b.getPieceAt(r, c);
        if(!this.equals(indicatedPiece)){
            throw new NotMeException("Attempting to move piece " +
              indicatedPiece.toString() + " using " + this.toString());
        }
        HashSet<Point> moveSet = new HashSet<Point>();
        return moveSet;
    }
}
